完成控件和Timeline的绘制

+++ Assets/AE_SkillEditor_Plus/Editor/UI/Data/TrackStyleData.cs
@@ -8,6 +8,6 @@ namespace AE_SkillEditor_Plus.UI.Data
         public string Name;
         public List<ClipStyleData> Clips;
         public Color Color;
-        public int WidthPreFrame;
+        public float WidthPreFrame;
     }
 }
